feat(biome-js-analyze): introduce class_member_analyzer to find all reads and writes of class members and props
