+++ crates/biome_js_analyze/src/lint/correctness/no_unused_private_class_members.rs
@@ -8,7 +8,7 @@ use biome_analyze::{
 use biome_console::markup;
 use biome_diagnostics::Severity;
 use biome_js_syntax::{
-    AnyJsClassMember, AnyJsClassMemberName, AnyJsFormalParameter, JsClassDeclaration,
+    AnyJsClassMember, AnyJsClassMemberName, AnyJsFormalParameter, JsClassDeclaration, JsSyntaxKind,
     TsAccessibilityModifier, TsPropertyParameter,
 };
 use biome_rowan::{
@@ -102,7 +102,11 @@ impl Rule for NoUnusedPrivateClassMembers {
 
                 let is_write_only = !is_read && is_write;
 
-                if is_write_only {
+                if !is_read && !is_write && private_member.is_accessor() {
+                    return Some(private_member.clone());
+                }
+
+                if is_write_only && private_member.is_accessor() {
                     return None;
                 }
 
@@ -178,6 +182,13 @@ fn get_constructor_params(class_declaration: &JsClassDeclaration) -> FxHashSet<A
 }
 
 impl AnyMember {
+    fn is_accessor(&self) -> bool {
+        matches!(
+            self.syntax().kind(),
+            JsSyntaxKind::JS_SETTER_CLASS_MEMBER | JsSyntaxKind::JS_GETTER_CLASS_MEMBER
+        )
+    }
+
     fn is_private(&self) -> bool {
         match self {
             Self::AnyJsClassMember(member) => {
